Fix: 배포 시 fetch error 해결
